Consolidate mini-swe integration under Slime
Remove the single-use common agent package and keep mini-swe protocol adaptation beside the Slime SWE agent while leaving benchmark submission parsing in SWE-Rebench.
