--- modal_training_gym/common/agents/__init__.py
@@ -1,5 +1,0 @@
-"""Framework-neutral agent protocol helpers."""
-
-from modal_training_gym.common.agents.mini_swe import MiniSweEnvironmentAdapter
-
-__all__ = ["MiniSweEnvironmentAdapter"]
